+++ src/features/ballot/components/BallotAllocationForm/index.ts
@@ -1,0 +1,1 @@
+export * from "./BallotAllocationForm";
